// cms/types/applyRole.ts
export interface IApplyRole {
  id: number;
  user_id: number;
  nickname: string;
  apply_type: string;
  company_name: string;
  email: string;
  contact_email: string;
  attach_file_name_1st: string;
  attach_file_name_2nd: string;
  attach_file_path_1st: string;
  attach_file_path_2nd: string;
  approval_code: string;
  approval_message: string;
  approval_date: string;
  createdDate: string;
  updatedDate: string;
}